1. Добавил удаление задач 2. Добавил звук который срабатывает при выполении задачи 3.  добавил стили для красной кнопки severity  = "danger" 4. Сделал чтобы описание задачи отображались как и в textarea 5. Установил alembic и обновил requirements.txt

+++ backend/tasks/service.py
@@ -88,10 +88,13 @@ class TaskService:
 
     
 
-    def task_delete(self,task_id:int):
+    def task_delete(self,task_id:int,current_user:User) -> Task:
         task = self.repo.get_task_by_id(task_id)
         if task is None : 
-            raise ValueError('Задача не найдена')
+            raise TaskNotFoundError('Задачи с таким id не существует')
+        
+        if task.author_id != current_user.id:
+            raise TaskIncorrectAuthorError('Это не ваша задача')
         
         return self.repo.delete_task(task)
